게시판 수정

+++ src/freeboard/service/DeleteBoardService.java
@@ -1,17 +1,30 @@
 package freeboard.service;
 
+import Exception.ArticleNullException;
+import Exception.UserNoNotMatchException;
 import freeboard.dao.FreeBoardDAO;
+import freeboard.model.FreeBoardList;
 
 public class DeleteBoardService {
 
 	
 	FreeBoardDAO freeBoardDAO = new FreeBoardDAO();
 	
-	public int delete(String no) {
+	public void delete(int no,int userNo) {
 		
+		FreeBoardList freeBoard = freeBoardDAO.getArticleNo(no);
 		
-		 int result = freeBoardDAO.deleteBoard(no);
+		if(freeBoard==null) {
+			throw new ArticleNullException();
+		}
+		if(canModify(userNo,freeBoard)) {
+			throw new UserNoNotMatchException();
+		}
 		
-		return result;
+		 freeBoardDAO.deleteBoard(no);
+		
 	}
+	private boolean canModify(int loginUserNo,FreeBoardList freeBoard) {
+		return loginUserNo!=freeBoard.getUserNo();
+}
 }
